AABB 구현 및 Camera 기능 추가

+++ D2DRender/Shape.cpp
@@ -34,8 +34,8 @@ void Shape::Render(D2DRenderer* _Renderer)
 		break;
 	case ShapeType::Circle:
 		D2D1_ELLIPSE a;
-		a.point.x = Position.x;
-		a.point.y = Position.y;
+		a.point.x = (float)Position.x;
+		a.point.y = (float)Position.y;
 		a.radiusX = 50.f;
 		a.radiusY = 50.f;
 		_Renderer->m_pD2DRenderTarget->DrawEllipse(a, _Renderer->m_pBlackBrush);
